Fixed a lot of issues with reader remotes migrating to WPKit
