+++ WordPress/Classes/Models/BasePost.h
@@ -34,15 +34,6 @@ NS_ASSUME_NONNULL_BEGIN
  */
 @property (nonatomic, assign) BOOL isFeaturedImageChanged;
 
-/**
- Create a summary for the post based on the post's content.
-
- @param string The post's content string. This should be the formatted content string.
- @return A summary for the post.
- */
-+ (NSString *)summaryFromContent:(NSString *)string;
-
-
 //date conversion
 @property (nonatomic, strong, nullable) NSDate * dateCreated;
