+++ WordPressKit/WordPressKit/RemoteReaderPost.m
@@ -1,6 +1,6 @@
 #import "RemoteReaderPost.h"
 #import "RemoteSourcePostAttribution.h"
-#import "WordPress-Swift.h"
+#import <WordPressKit/WordPressKit-Swift.h>
 
 @implementation RemoteReaderPost
